Move wallet functions from bitcoinrpc.cpp to rpcwallet.cpp
Moves all rpc externs to bitcoinrpc.h
Adds JSONRPCError prototype to bitcoinrpc.h (vs externing where required)
Removes unused namespace declarations from rpcrawtransaction.cpp
