+++ src/makefile.linux-mingw
@@ -62,6 +62,7 @@ OBJS= \
     obj/bitcoinrpc.o \
     obj/rpcdump.o \
     obj/rpcnet.o \
+    obj/rpcwallet.o \
     obj/rpcrawtransaction.o \
     obj/script.o \
     obj/util.o \
